add github-bot-madgerc.js
