+++ .madgerc
@@ -1,0 +1,3 @@
+{
+    "tsConfig": "tsconfig.json"
+}
