+++ github-bot-madgerc.js
@@ -1,0 +1,6 @@
+const path = require('path');
+
+module.exports = {
+  src: path.resolve(__dirname, './src/index.tsx'),
+  tsConfig: path.resolve(__dirname, './tsconfig.json'),
+};
